Add-per-dojo-solve-count-to-dojos-page
