+++ dojo_plugin/utils/__init__.py
@@ -27,71 +27,6 @@ from sqlalchemy import String, Integer
 from sqlalchemy.sql import or_
 
 
-PLUGIN_DIR = pathlib.Path(__file__).parent.parent
-CHALLENGES_DIR = pathlib.Path("/var/challenges")
-DOJOS_DIR = pathlib.Path("/var/dojos")
-DATA_DIR = pathlib.Path("/var/data")
-
-INDEX_HTML = pathlib.Path("/var/index.html").read_text()
-
-def create_seccomp():
-    seccomp = json.load(pathlib.Path("/etc/docker/seccomp.json").open())
-
-    seccomp["syscalls"].append({
-        "names": [
-            "clone",
-            "sethostname",
-            "setns",
-            "unshare",
-        ],
-        "action": "SCMP_ACT_ALLOW",
-    })
-
-    READ_IMPLIES_EXEC = 0x0400000
-    ADDR_NO_RANDOMIZE = 0x0040000
-
-    existing_personality_values = []
-    for syscalls in seccomp["syscalls"]:
-        if "personality" not in syscalls["names"]:
-            continue
-        if syscalls["action"] != "SCMP_ACT_ALLOW":
-            continue
-        assert len(syscalls["args"]) == 1
-        arg = syscalls["args"][0]
-        assert list(arg.keys()) == ["index", "value", "op"]
-        assert arg["index"] == 0, arg
-        assert arg["op"] == "SCMP_CMP_EQ"
-        existing_personality_values.append(arg["value"])
-
-    new_personality_values = []
-    for new_flag in [READ_IMPLIES_EXEC, ADDR_NO_RANDOMIZE]:
-        for value in [0, *existing_personality_values]:
-            new_value = value | new_flag
-            if new_value not in existing_personality_values:
-                new_personality_values.append(new_value)
-                existing_personality_values.append(new_value)
-
-    for new_value in new_personality_values:
-        seccomp["syscalls"].append({
-            "names": ["personality"],
-            "action": "SCMP_ACT_ALLOW",
-            "args": [
-                {
-                    "index": 0,
-                    "value": new_value,
-                    "op": "SCMP_CMP_EQ",
-                },
-            ],
-        })
-
-    return json.dumps(seccomp)
-SECCOMP = create_seccomp()
-
-USER_FIREWALL_ALLOWED = {
-    host: socket.gethostbyname(host)
-    for host in pathlib.Path("/var/user_firewall.allowed").read_text().split()
-}
-
 ID_REGEX = "^[A-Za-z0-9_.-]+$"
 def id_regex(s):
     return re.match(ID_REGEX, s) and ".." not in s
